move to util to avoid consensus depending on wallet
