+++ chia/_tests/core/full_node/stores/test_block_store.py
@@ -31,7 +31,7 @@ from chia.types.blockchain_format.serialized_program import SerializedProgram
 from chia.types.blockchain_format.vdf import VDFProof
 from chia.util.db_wrapper import get_host_parameter_limit
 from chia.util.task_referencer import create_referenced_task
-from chia.wallet.util.casts import int_to_bytes
+from chia.util.casts import int_to_bytes
 
 log = logging.getLogger(__name__)
