+++ chia/_tests/core/mempool/test_mempool.py
@@ -76,7 +76,7 @@ from chia.util.hash import std_hash
 from chia.util.recursive_replace import recursive_replace
 from chia.wallet.conditions import AssertCoinAnnouncement, AssertPuzzleAnnouncement
 from chia.wallet.estimate_fees import estimate_fees
-from chia.wallet.util.casts import int_to_bytes
+from chia.util.casts import int_to_bytes
 
 BURN_PUZZLE_HASH = bytes32(b"0" * 32)
 BURN_PUZZLE_HASH_2 = bytes32(b"1" * 32)
