+++ chia/consensus/condition_tools.py
@@ -14,7 +14,7 @@ from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.errors import ConsensusError, Err
 from chia.util.hash import std_hash
-from chia.wallet.util.casts import int_from_bytes, int_to_bytes
+from chia.util.casts import int_from_bytes, int_to_bytes
 
 
 def parse_sexp_to_condition(sexp: Program) -> ConditionWithArgs:
