+++ chia/pools/pool_puzzles.py
@@ -28,7 +28,7 @@ from chia.wallet.puzzles.singleton_top_layer import (
     SINGLETON_MOD_HASH,
     puzzle_for_singleton,
 )
-from chia.wallet.util.casts import int_to_bytes
+from chia.util.casts import int_to_bytes
 from chia.wallet.util.compute_additions import compute_additions
 from chia.wallet.util.curry_and_treehash import calculate_hash_of_quoted_mod_hash, curry_and_treehash, shatree_atom
