+++ chia/wallet/util/compute_memos.py
@@ -7,7 +7,7 @@ from chia_rs.sized_ints import uint64
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.program import INFINITE_COST, run_with_cost
 from chia.types.condition_opcodes import ConditionOpcode
-from chia.wallet.util.casts import int_from_bytes
+from chia.util.casts import int_from_bytes
 from chia.wallet.wallet_spend_bundle import WalletSpendBundle
