+++ chia/wallet/util/merkle_tree.py
@@ -7,7 +7,7 @@ from typing import Optional
 from chia_rs.sized_bytes import bytes32
 
 from chia.util.hash import std_hash
-from chia.wallet.util.casts import int_to_bytes
+from chia.util.casts import int_to_bytes
 
 ONE = int_to_bytes(1)
 TWO = int_to_bytes(2)
