+++ chia/wallet/wallet_singleton_store.py
@@ -20,7 +20,7 @@ from chia.wallet import singleton
 from chia.wallet.lineage_proof import LineageProof
 from chia.wallet.singleton import get_inner_puzzle_from_singleton, get_singleton_id_from_puzzle
 from chia.wallet.singleton_record import SingletonRecord
-from chia.wallet.util.casts import int_from_bytes
+from chia.util.casts import int_from_bytes
 
 log = logging.getLogger(__name__)
